+++ Veterinaria.Web/Controllers/PetTypesController.cs
@@ -123,28 +123,22 @@ namespace Veterinaria.Web.Controllers
             {
                 return NotFound();
             }
-
             var petType = await _context.PetTypes
+                .Include(pt => pt.Pets)
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (petType == null)
             {
                 return NotFound();
             }
-
-            return View(petType);
-        }
-
-        // POST: PetTypes/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
-        {
-            var petType = await _context.PetTypes.FindAsync(id);
+            if (petType.Pets.Count > 0)
+            {
+                ModelState.AddModelError(string.Empty, "El tipo de mascota tiene registros relacionados");
+                return RedirectToAction(nameof(Index));
+            }
             _context.PetTypes.Remove(petType);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
-
         private bool PetTypeExists(int id)
         {
             return _context.PetTypes.Any(e => e.Id == id);
